generalize xhr request in utils for resuse
