+++ public/js/bookmark-buttons.js
@@ -1,3 +1,5 @@
+import { xhrReq } from "./utils/utils.js";
+
 const bookmarkButtons = {
   init() {
     const bookmarkButtons = Array.prototype.slice.call(
@@ -15,23 +17,16 @@ const bookmarkButtons = {
   },
 
   toggleBookmark(addOrDelete, linkEl, isBookmarked, thingid, bookmarkType) {
-    const errorCodes = [500, 400, 401];
-    const request = new XMLHttpRequest();
     let action = "POST";
     if (addOrDelete === "delete") {
       action = "DELETE";
     }
-    request.open(action, `/bookmark/${addOrDelete}`, true);
-    request.onreadystatechange = () => {
-      if (request.readyState === 4 && errorCodes.includes(request.status)) {
-        console.error("error: " + request.responseText)
-      } else if (request.readyState === 4) {
-        // success, toggle data attribute to update icon style
-        linkEl.setAttribute("data-bookmarked", !isBookmarked);
-      }
-    };
-    request.setRequestHeader('Content-Type', 'application/json')
-    request.send(JSON.stringify({ thingid, bookmarkType }));
+    const successCB = () => {
+      linkEl.setAttribute("data-bookmarked", !isBookmarked);
+    }
+    const data = JSON.stringify({ thingid, bookmarkType });
+    const url = `/bookmark/${addOrDelete}`;
+    xhrReq(action, url, data, successCB);
   },
 
   handleLinkClick(e) {
